Review rework, make A2A fail if it is not installed
